Mod updates & more testing

+++ py/delete_files.py
@@ -15,15 +15,16 @@ for file_name in files_to_delete:
     file_path = os.path.join(destination_path, file_name)
     try:
         if file_name == 'HoT-CatModLoader.exe':
-            print(f'Delete ["{file_name}"]')
-            user_input = input(f'\n[Y/N]: ')
-            if user_input == 'y' or user_input == 'Y':
-                try:
-                    os.remove(file_path)
-                except Exception as error:
-                    print(str(error))
-            else:
-                exit()
+            if os.path.exists(file_path):
+                print(f'Delete ["{file_name}"]')
+                user_input = input(f'\n[Y/N]: ')
+                if user_input == 'y' or user_input == 'Y':
+                    try:
+                        os.remove(file_path)
+                    except Exception as error:
+                        print(str(error))
+                else:
+                    exit()
         os.remove(file_path)
         print(f"File '{file_name}' deleted successfully.")
     except FileNotFoundError:
